refactoring (sorry for bad commit messages)

+++ core/src/motor_control/motor_drivers/l298n_motor_driver.cpp
@@ -1,4 +1,5 @@
 #include "motor_control/motor_drivers/l298n_motor_driver.hpp"
+#include <Arduino.h>
 
 L298NMotorDriver::L298NMotorDriver(unsigned int pin_in1, unsigned int pin_in2, unsigned int pin_ena, unsigned int pwm_channel)
   : pin_in1_(pin_in1), pin_in2_(pin_in2), pin_ena_(pin_ena), pwm_channel_(pwm_channel) {
@@ -25,6 +26,6 @@ void L298NMotorDriver::set_motor_control(float control_value) {
   digitalWrite(pin_in2_, direction ? LOW : HIGH);
 
   // Set PWM for L298N...
-  int pwm = static_cast<int>(std::abs(control_value) * 255);
+  u_int8_t pwm = static_cast<int>(std::abs(control_value) * 255);
   ledcWrite(pwm_channel_, pwm);
 }
